server/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
sqlmodel>=0.0.22
asyncpg>=0.30
sqlalchemy[asyncio]>=2.0
alembic>=1.14
pydantic-settings>=2.7
python-multipart>=0.0.18
Pillow>=11.0
aiofiles>=24.1
